packages/virtual-assistant/virtual-assistant-livekit/pyproject.toml
[project]
name = "virtual-assistant-livekit"
version = "1.0.0"
description = "LiveKit integration for Virtual Assistant with Amazon Nova Sonic"
requires-python = ">=3.13,<4"
readme = "README.md"
dependencies = [
    # Official PyPI release with Nova Sonic 2 support
    "livekit-agents[mcp]>=1.3.9",
    "livekit-plugins-aws[realtime]>=1.3.9",
    "aws_sdk_bedrock_runtime",
    "httpx>=0.25.0",
    "boto3>=1.34.0",
    "virtual-assistant-common",
]

[tool.hatch.build.targets.wheel]
packages = ["virtual_assistant_livekit"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "autopep8>=2.3.1",
    "ruff>=0.8.2",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "moto[secretsmanager]>=5.0.16",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "dist",
]

line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]
ignore = []

fixable = ["ALL"]
unfixable = []

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests",
]
testpaths = ["tests"]
addopts = "-m 'not integration'"
asyncio_mode = "auto"
